break down logic into packages and separate get and update services
