--- handlers/delivery-records-api/src/main/scala/com/gu/delivery_records_api/DeliveryRecordsApiResponse.scala
@@ -1,7 +1,0 @@
-package com.gu.delivery_records_api
-
-case class DeliveryRecordsApiResponse(
-    results: List[DeliveryRecord],
-    deliveryProblemMap: Map[String, DeliveryProblemCase],
-    contactPhoneNumbers: SFApiContactPhoneNumbers,
-)
