--- handlers/delivery-records-api/src/main/scala/com/gu/delivery_records_api/SFApiSubscription.scala
@@ -1,8 +1,0 @@
-package com.gu.delivery_records_api
-
-import com.gu.salesforce.RecordsWrapperCaseClass
-
-case class SFApiSubscription(
-    Buyer__r: SFApiContactPhoneNumbers,
-    Delivery_Records__r: Option[RecordsWrapperCaseClass[SFApiDeliveryRecord]],
-)
