+++ handlers/delivery-records-api/src/main/scala/com/gu/delivery_records_api/service/DeliveryRecordsServiceError.scala
@@ -1,0 +1,11 @@
+package com.gu.delivery_records_api.service
+
+sealed trait DeliveryRecordServiceError
+
+case class DeliveryRecordServiceGenericError(message: String) extends DeliveryRecordServiceError
+
+case class DeliveryRecordServiceSubscriptionNotFound(message: String) extends DeliveryRecordServiceError
+
+
+
+
