+++ handlers/delivery-records-api/src/main/scala/com/gu/delivery_records_api/service/createproblem/SFApiDeliveryRecord.scala
@@ -1,4 +1,4 @@
-package com.gu.delivery_records_api
+package com.gu.delivery_records_api.service.createproblem
 
 import java.time.LocalDate
