+++ handlers/delivery-records-api/src/test/scala/com/gu/delivery_records_api/DeliveryRecordsApiTest.scala
@@ -1,13 +1,15 @@
 package com.gu.delivery_records_api
 
-import cats.effect.IO
-import com.gu.delivery_records_api.DeliveryRecordsServiceImpl.deliveryRecordsQuery
+import cats.effect.{ContextShift, IO}
+import com.gu.delivery_records_api.service.createproblem._
+import com.gu.delivery_records_api.service.getrecords.GetDeliveryRecordsServiceImpl.deliveryRecordsQuery
+import com.gu.delivery_records_api.service.getrecords.{DeliveryProblemCase, DeliveryProblemCredit, DeliveryRecord, DeliveryRecordsApiResponse}
 import sttp.client3.impl.cats.CatsMonadAsyncError
 import sttp.client3.testing.SttpBackendStub
 import com.gu.salesforce._
 import com.gu.salesforce.sttp.SalesforceStub.implicitSalesforceStub
 import com.gu.salesforce.sttp.{SFApiCompositeResponse, SFApiCompositeResponsePart}
-import io.circe.Decoder
+import io.circe.{Decoder, Json}
 import io.circe.generic.auto._
 import io.circe.parser._
 import io.circe.syntax._
@@ -22,11 +24,11 @@ import scala.concurrent._
 
 class DeliveryRecordsApiTest extends AnyFlatSpec with Matchers with EitherValues {
 
-  private implicit val contextShift = IO.contextShift(ExecutionContext.global)
+  private implicit val contextShift: ContextShift[IO] = IO.contextShift(ExecutionContext.global)
 
-  val config =
+  val config: SFAuthConfig =
     SFAuthConfig("https://salesforceAuthUrl", "sfClientId", "sfClientSecret", "sfUsername", "sfPassword", "sfToken")
-  val auth = SalesforceAuth("salesforce-access-token", "https://salesforceInstanceUrl")
+  val auth: SalesforceAuth = SalesforceAuth("salesforce-access-token", "https://salesforceInstanceUrl")
   val subscriptionNumber = "A-213123"
   val identityId = "identity id"
   val buyerContactId = "contact id"
@@ -45,7 +47,7 @@ class DeliveryRecordsApiTest extends AnyFlatSpec with Matchers with EitherValues
   val hasHolidayStop = true
   val doesntHaveHolidayStop = false
   val bulkSuspensionReason = "Covid-19"
-  val sfProblemCase = SFApiDeliveryProblemCase(
+  val sfProblemCase: SFApiDeliveryProblemCase = SFApiDeliveryProblemCase(
     Id = "case_id",
     CaseNumber = "123456",
     Subject = Some("subject"),
@@ -53,9 +55,9 @@ class DeliveryRecordsApiTest extends AnyFlatSpec with Matchers with EitherValues
     Case_Closure_Reason__c = Some("Paper Damaged"),
   )
   val creditAmount = 1.23
-  val invoiceDate = LocalDate.of(2019, 12, 10)
+  val invoiceDate: LocalDate = LocalDate.of(2019, 12, 10)
   val isActioned = true
-  val contactNumbers = SFApiContactPhoneNumbers(
+  val contactNumbers: SFApiContactPhoneNumbers = SFApiContactPhoneNumbers(
     Id = Some("id"),
     Phone = Some("+447654321234"),
     HomePhone = Some("+441234567890"),
@@ -63,7 +65,7 @@ class DeliveryRecordsApiTest extends AnyFlatSpec with Matchers with EitherValues
     OtherPhone = None,
   )
 
-  val sfDeliveryRecordA = SFApiDeliveryRecord(
+  val sfDeliveryRecordA: SFApiDeliveryRecord = SFApiDeliveryRecord(
     Id = deliveryRecordId,
     Delivery_Date__c = Some(deliveryDate),
     Delivery_Address__c = Some(deliveryAddress1),
@@ -94,7 +96,7 @@ class DeliveryRecordsApiTest extends AnyFlatSpec with Matchers with EitherValues
     Delivery_Instructions__c = Some(deliveryInstruction2),
   )
 
-  val validSalesforceResponseBody = RecordsWrapperCaseClass(
+  val validSalesforceResponseBody: RecordsWrapperCaseClass[SFApiSubscription] = RecordsWrapperCaseClass(
     List(
       SFApiSubscription(
         Buyer__r = contactNumbers,
@@ -113,7 +115,7 @@ class DeliveryRecordsApiTest extends AnyFlatSpec with Matchers with EitherValues
     ),
   )
 
-  val expectedDeliveryRecordA = DeliveryRecord(
+  val expectedDeliveryRecordA: DeliveryRecord = DeliveryRecord(
     id = deliveryRecordId,
     deliveryDate = Some(deliveryDate),
     deliveryInstruction = Some(deliveryInstruction1),
@@ -152,7 +154,7 @@ class DeliveryRecordsApiTest extends AnyFlatSpec with Matchers with EitherValues
     deliveryAddress = Some(deliveryAddress2),
   )
 
-  val expectedValidDeliveryApiResponse = DeliveryRecordsApiResponse(
+  val expectedValidDeliveryApiResponse: DeliveryRecordsApiResponse = DeliveryRecordsApiResponse(
     List(
       expectedDeliveryRecordB,
       expectedDeliveryRecordB.copy(
@@ -175,7 +177,7 @@ class DeliveryRecordsApiTest extends AnyFlatSpec with Matchers with EitherValues
     contactNumbers.copy(MobilePhone = None),
   )
 
-  val validCompositeResponse = SFApiCompositeResponse(
+  val validCompositeResponse: SFApiCompositeResponse = SFApiCompositeResponse(
     List(
       SFApiCompositeResponsePart(200, "CaseCreation"),
       SFApiCompositeResponsePart(200, "LinkDeliveryRecord-deliveryRecordID"),
@@ -183,7 +185,7 @@ class DeliveryRecordsApiTest extends AnyFlatSpec with Matchers with EitherValues
     ),
   )
 
-  val failedCompositeResponse = SFApiCompositeResponse(
+  val failedCompositeResponse: SFApiCompositeResponse = SFApiCompositeResponse(
     List(
       SFApiCompositeResponsePart(400, "CaseCreation"),
       SFApiCompositeResponsePart(400, "LinkDeliveryRecord-deliveryRecordID"),
@@ -206,7 +208,7 @@ class DeliveryRecordsApiTest extends AnyFlatSpec with Matchers with EitherValues
       .run(
         Request(
           method = Method.GET,
-          Uri(path = s"/delivery-records/${subscriptionNumber}"),
+          Uri(path = s"/delivery-records/$subscriptionNumber"),
           headers = Headers.of(Header("x-identity-id", identityId)),
         ),
       )
@@ -233,7 +235,7 @@ class DeliveryRecordsApiTest extends AnyFlatSpec with Matchers with EitherValues
       .run(
         Request(
           method = Method.GET,
-          Uri(path = s"/delivery-records/${subscriptionNumber}"),
+          Uri(path = s"/delivery-records/$subscriptionNumber"),
           headers = Headers.of(Header("x-salesforce-contact-id", buyerContactId)),
         ),
       )
@@ -269,7 +271,7 @@ class DeliveryRecordsApiTest extends AnyFlatSpec with Matchers with EitherValues
         Request(
           method = Method.GET,
           Uri(
-            path = s"/delivery-records/${subscriptionNumber}",
+            path = s"/delivery-records/$subscriptionNumber",
             query = Query("startDate" -> Some(startDate.toString), "endDate" -> Some(endDate.toString)),
           ),
           headers = Headers.of(Header("x-salesforce-contact-id", buyerContactId)),
@@ -297,7 +299,7 @@ class DeliveryRecordsApiTest extends AnyFlatSpec with Matchers with EitherValues
       .run(
         Request(
           method = Method.GET,
-          Uri(path = s"/delivery-records/${subscriptionNumber}"),
+          Uri(path = s"/delivery-records/$subscriptionNumber"),
           headers = Headers.of(Header("x-salesforce-contact-id", buyerContactId)),
         ),
       )
@@ -308,7 +310,7 @@ class DeliveryRecordsApiTest extends AnyFlatSpec with Matchers with EitherValues
     response.status.code should equal(404)
   }
 
-  val createDeliveryProblemBody = CreateDeliveryProblem(
+  val createDeliveryProblemBody: Json = CreateDeliveryProblem(
     productName = "Guardian Weekly",
     description = Some("String"),
     problemType = "No Delivery",
@@ -344,7 +346,7 @@ class DeliveryRecordsApiTest extends AnyFlatSpec with Matchers with EitherValues
       .run(
         Request(
           method = Method.POST,
-          Uri(path = s"/delivery-records/${subscriptionNumber}"),
+          Uri(path = s"/delivery-records/$subscriptionNumber"),
           headers = Headers.of(Header("x-identity-id", identityId)),
         ).withEntity(
           createDeliveryProblemBody,
@@ -374,7 +376,7 @@ class DeliveryRecordsApiTest extends AnyFlatSpec with Matchers with EitherValues
       .run(
         Request(
           method = Method.GET,
-          Uri(path = s"/delivery-records/${subscriptionNumber}"),
+          Uri(path = s"/delivery-records/$subscriptionNumber"),
           headers = Headers.of(Header("x-salesforce-contact-id", buyerContactId)),
         ).withEntity(
           createDeliveryProblemBody,
@@ -404,7 +406,7 @@ class DeliveryRecordsApiTest extends AnyFlatSpec with Matchers with EitherValues
       .run(
         Request(
           method = Method.POST,
-          Uri(path = s"/delivery-records/${subscriptionNumber}"),
+          Uri(path = s"/delivery-records/$subscriptionNumber"),
           headers = Headers.of(Header("x-identity-id", identityId)),
         ).withEntity(
           createDeliveryProblemBody,
@@ -427,7 +429,7 @@ class DeliveryRecordsApiTest extends AnyFlatSpec with Matchers with EitherValues
       .run(
         Request(
           method = Method.GET,
-          Uri(path = s"/delivery-records/${subscriptionNumber}"),
+          Uri(path = s"/delivery-records/$subscriptionNumber"),
           headers = Headers.of(Header("x-salesforce-contact-id", buyerContactId)),
         ),
       )
@@ -448,7 +450,7 @@ class DeliveryRecordsApiTest extends AnyFlatSpec with Matchers with EitherValues
       .run(
         Request(
           method = Method.GET,
-          Uri(path = s"/delivery-records/${subscriptionNumber}"),
+          Uri(path = s"/delivery-records/$subscriptionNumber"),
         ),
       )
       .value
@@ -481,3 +483,8 @@ class DeliveryRecordsApiTest extends AnyFlatSpec with Matchers with EitherValues
     }
   }
 }
+
+case class SFApiSubscription(
+  Buyer__r: SFApiContactPhoneNumbers,
+  Delivery_Records__r: Option[RecordsWrapperCaseClass[SFApiDeliveryRecord]],
+)
